DUmmy branching diameters are assigned based on mean values for the associate vascular tree - lengths are now assigned to 5 um (simulations are sensitive to this value). r0 is now assigned based on the half-distance to the nearest node (avoids overlap). Preloaded cross micro-cells have been corrected for user applies rotation. 2D micro-cells are assigned a z-depth = to max diameter. Error when computing micro-cell conductance has been corrected. Flow solver for micro-cell problem has been updated and simplified - corrections have been made for off-axis vessel centrelines. General housekeeping.

+++ Reanimate/MicroCell.cpp
@@ -16,6 +16,7 @@ void MicroCell::setup_microCellArrays()  {
 
     nodsegm = 10;
     computeLseg = 1;
+
     setup_networkArrays();
 
     segname = zeros<ivec>(nseg);
@@ -38,8 +39,8 @@ void MicroCell::setup_mcFlowArrays() {
     c = zeros<vec>(nseg);
 
     cA = zeros<mat>(nseg,nnod);
-    cC = zeros<mat>(nnodbc/2.,nnod);
-    cE = zeros<mat>(nnodbc/2.,nseg);
+    cC = zeros<mat>(int(nnodbc/2.),nnod);
+    cE = zeros<mat>(int(nnodbc/2.),nseg);
     cF = zeros<mat>(nseg,nnod);
     conductivity = zeros<mat>(3,3);
